requirements.txt
anthropic>=0.40.0
typer[all]>=0.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
python-dateutil>=2.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
